# swagger.yaml
openapi: 3.0.3
info:
  title: CPRT LIS API
  version: 1.0.0
  description: Full API specification for all currently registered endpoints.
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
  - name: Auth
  - name: Users
  - name: Patients
  - name: Billing
  - name: Orders
  - name: Lab

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              byLoginId:
                value:
                  login_id: superadmin
                  password: password123
              byUsername:
                value:
                  username: superadmin
                  password: password123
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoginSuccess'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users:
    get:
      tags: [Users]
      summary: List all users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'

  /api/users/{userId}:
    patch:
      tags: [Users]
      summary: Update user account
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpdateUserResponse'

  /api/patients/register:
    post:
      tags: [Patients]
      summary: Register patient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient registered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientRegisterResponse'

  /api/patients:
    post:
      tags: [Patients]
      summary: Create patient (alias of register)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientRegisterResponse'
    get:
      tags: [Patients]
      summary: Search patients
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Patient search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSearchResult'

  /api/patients/search:
    get:
      tags: [Patients]
      summary: Search patients (explicit route)
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Patient search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSearchResult'

  /api/patients/{id}:
    get:
      tags: [Patients]
      summary: Get patient history by patient ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientHistoryItem'

  /api/patients/{patientId}/history:
    get:
      tags: [Patients]
      summary: Get patient history
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientHistoryItem'

  /api/patients/{patientId}:
    patch:
      tags: [Patients]
      summary: Update patient profile
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientProfile'

  /api/billing/new:
    post:
      tags: [Billing]
      summary: Generate bill
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateBillRequest'
      responses:
        '201':
          description: Bill generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenerateBillResponse'

  /api/billing/bills:
    post:
      tags: [Billing]
      summary: Create bill (legacy/internal style)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBillRequest'
      responses:
        '201':
          description: Bill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabBill'

  /api/billing/bills/{id}/items:
    post:
      tags: [Billing]
      summary: Add item to bill
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBillItemRequest'
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: added

  /api/billing/services:
    get:
      tags: [Billing]
      summary: List billable services
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabService'

  /api/billing/{billId}/payment:
    patch:
      tags: [Billing]
      summary: Update payment for bill
      security:
        - BearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaymentRequest'
      responses:
        '200':
          description: Payment updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BillPaymentUpdate'

  /api/orders:
    post:
      tags: [Orders]
      summary: Create lab order
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabOrder'

  /api/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated

  /api/lab/sample-collection/{billId}:
    patch:
      tags: [Lab]
      summary: Mark sample collection
      security:
        - BearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SampleCollectionRequest'
      responses:
        '200':
          description: Sample collection marked
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SampleCollectionResponse'

  /api/lab/results/verify:
    post:
      tags: [Lab]
      summary: Verify lab results
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyResultsRequest'
      responses:
        '200':
          description: Results verified
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultVerificationResponse'

  /api/lab/results/certify/{billId}:
    patch:
      tags: [Lab]
      summary: Certify lab results
      description: Restricted by middleware policy to DOCTOR role.
      security:
        - BearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertifyResultsRequest'
      responses:
        '200':
          description: Results certified
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultCertificationResponse'

  /api/lab/reports/{billId}:
    get:
      tags: [Lab]
      summary: Get lab report
      security:
        - BearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Report fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabReportResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok

    LoginRequest:
      type: object
      properties:
        login_id:
          type: string
        username:
          type: string
        password:
          type: string
      required: [password]

    LoginUser:
      type: object
      properties:
        user_id:
          type: string
        user_name:
          type: string
        account_group_code:
          type: string
        account_group_name:
          type: string
        role:
          type: string
        status:
          type: string

    LoginSuccess:
      type: object
      properties:
        status:
          type: string
        token:
          type: string
        user:
          $ref: '#/components/schemas/LoginUser'
        permissions:
          type: array
          items:
            type: string

    UserSummary:
      type: object
      properties:
        user_id:
          type: string
        login_id:
          type: string
        user_name:
          type: string
        account_group:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true

    UpdateUserRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Active, Inactive]
        account_group_id:
          type: integer
          format: int64
        updated_by:
          type: string
        password:
          type: string
        password_hash:
          type: string

    UpdateUserResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        user_id:
          type: string
        new_status:
          type: string

    CreatePatientRequest:
      type: object
      properties:
        prefix:
          type: string
        first_name:
          type: string
        gender:
          type: string
        age:
          type: integer
        age_unit:
          type: string
          enum: [Yrs, Mon, Days]
        phone_no:
          type: string
        phone:
          type: string
        op_ip_no:
          type: string
        patient_type:
          type: string
        created_by:
          type: string
        status:
          type: string
      required: [first_name, age]

    PatientRegisterResponse:
      type: object
      properties:
        patient_id:
          type: string
        patient_no:
          type: integer
          format: int64
        full_name:
          type: string
        registration_date:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time

    PatientSearchResult:
      type: object
      properties:
        patient_id:
          type: string
        full_name:
          type: string
        op_ip_no:
          type: string
        phone_no:
          type: string
        created_at:
          type: string
        created_by:
          type: string
        patient_no:
          type: integer
          format: int64
        patient_type:
          type: string
        status:
          type: string

    PatientHistoryItem:
      type: object
      properties:
        bill_date:
          type: string
        bill_no:
          type: integer
          format: int64
        service_name:
          type: string
        status:
          type: string
        rate:
          type: number
          format: double

    UpdatePatientRequest:
      type: object
      properties:
        prefix:
          type: string
        first_name:
          type: string
        gender:
          type: string
        age:
          type: integer
        age_unit:
          type: string
          enum: [Yrs, Mon, Days]
        phone_no:
          type: string
        patient_type:
          type: string
        status:
          type: string
        updated_by:
          type: string

    PatientProfile:
      type: object
      properties:
        patient_id:
          type: string
        prefix:
          type: string
        first_name:
          type: string
        gender:
          type: string
        age:
          type: integer
        age_unit:
          type: string
        phone_no:
          type: string
        patient_type:
          type: string
        status:
          type: string
        updated_by:
          type: string

    GenerateBillRequest:
      type: object
      properties:
        patient_id:
          type: string
        referred_by:
          type: string
        hospital_name:
          type: string
        services:
          type: array
          items:
            type: object
            properties:
              service_id:
                type: string
              service_name:
                type: string
              rate:
                type: number
                format: double
            required: [service_id, service_name, rate]
        discount_amt:
          type: number
          format: double
        tax_percent:
          type: number
          format: double
        received_amt:
          type: number
          format: double
      required: [patient_id, services]

    GenerateBillResponse:
      type: object
      properties:
        bill_id:
          type: string
        bill_no:
          type: integer
          format: int64
        total_billed_amt:
          type: number
          format: double
        discount_amt:
          type: number
          format: double
        tax_amt:
          type: number
          format: double
        net_billed_amt:
          type: number
          format: double
        received_amt:
          type: number
          format: double
        balance_amt:
          type: number
          format: double
        payment_status:
          type: string
        created_at:
          type: string

    CreateBillRequest:
      type: object
      properties:
        patient_id:
          type: integer
          format: int64
        visit_id:
          type: integer
          format: int64
          nullable: true
        doctor_id:
          type: integer
          format: int64
          nullable: true
        total_amount:
          type: number
          format: double
        discount:
          type: number
          format: double
        tax:
          type: number
          format: double
        net_amount:
          type: number
          format: double
        status:
          type: string
        payment_mode:
          type: string

    LabBill:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        BillUUID:
          type: string
        BillNo:
          type: integer
          format: int64
        PatientID:
          type: integer
          format: int64
        VisitID:
          type: integer
          format: int64
          nullable: true
        DoctorID:
          type: integer
          format: int64
          nullable: true
        TotalAmount:
          type: number
          format: double
        Discount:
          type: number
          format: double
        Tax:
          type: number
          format: double
        NetAmount:
          type: number
          format: double
        Status:
          type: string
        PaymentMode:
          type: string
        CreatedAt:
          type: string
          format: date-time

    AddBillItemRequest:
      type: object
      properties:
        service_id:
          type: integer
          format: int64
        qty:
          type: integer
        unit_price:
          type: number
          format: double
      required: [service_id, qty, unit_price]

    LabService:
      type: object
      properties:
        service_id:
          type: integer
          format: int64
        service_name:
          type: string
        rate:
          type: number
          format: double
        department:
          type: string
        status:
          type: string

    UpdatePaymentRequest:
      type: object
      properties:
        received_amt:
          type: number
          format: double
        payment_mode:
          type: string

    BillPaymentUpdate:
      type: object
      properties:
        bill_id:
          type: string
        bill_no:
          type: integer
          format: int64
        net_billed_amt:
          type: number
          format: double
        received_amt:
          type: number
          format: double
        balance_amt:
          type: number
          format: double
        payment_mode:
          type: string
        payment_status:
          type: string
        updated_at:
          type: string

    CreateOrderRequest:
      type: object
      properties:
        bill_id:
          type: integer
          format: int64
        patient_id:
          type: integer
          format: int64
        visit_id:
          type: integer
          format: int64
          nullable: true
        status:
          type: string
      required: [bill_id, patient_id]

    LabOrder:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        OrderUUID:
          type: string
        BillID:
          type: integer
          format: int64
        PatientID:
          type: integer
          format: int64
        VisitID:
          type: integer
          format: int64
          nullable: true
        Status:
          type: string
        CreatedAt:
          type: string
          format: date-time

    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SampleCollectionRequest:
      type: object
      properties:
        sample_no:
          type: string
        collected_by:
          type: string
        worksheet_no:
          type: string
      required: [sample_no, collected_by]

    SampleCollectionResponse:
      type: object
      properties:
        bill_id:
          type: string
        sample_no:
          type: string
        collection_status:
          type: string
        collected_by:
          type: string
        worksheet_no:
          type: string
        collection_datetime:
          type: string

    VerifyResultsRequest:
      type: object
      properties:
        bill_id:
          type: integer
          format: int64
        params:
          type: array
          items:
            $ref: '#/components/schemas/ResultVerificationParam'
        verified_by:
          type: string
      required: [bill_id, params, verified_by]

    ResultVerificationParam:
      type: object
      properties:
        param_id:
          type: string
        param_name:
          type: string
        result_value:
          type: string
        is_abnormal:
          type: boolean
      required: [param_id, param_name, result_value, is_abnormal]

    ResultVerificationResponse:
      type: object
      properties:
        bill_id:
          type: string
        verification_status:
          type: string
        verified_by:
          type: string
        verification_datetime:
          type: string
        result_count:
          type: integer

    CertifyResultsRequest:
      type: object
      properties:
        certified_by:
          type: string
        remarks:
          type: string
      required: [certified_by]

    ResultCertificationResponse:
      type: object
      properties:
        bill_id:
          type: string
        certification_status:
          type: string
        certified_by:
          type: string
        certification_remarks:
          type: string
        dispatch_ready:
          type: boolean

    LabReportResult:
      type: object
      properties:
        param_name:
          type: string
        result_value:
          type: string
        reference:
          type: string
        flag:
          type: string

    LabReportResponse:
      type: object
      properties:
        bill_id:
          type: string
        patient_id:
          type: string
        patient_name:
          type: string
        verification_by:
          type: string
        certified_by:
          type: string
        report_status:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/LabReportResult'